Modalに部位をセレクトできるボタンを追加しました
